LayoutTests/platform/ios/imported/w3c/web-platform-tests/uievents/mouse/mouseover-at-removing-mousedown-target_duration=16-expected.txt


FAIL mouseover should be fired before mouseup if mousedown target is removed (0) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (0) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (1) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (1) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (2) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (2) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (3) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (3) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (4) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (4) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (5) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (5) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (6) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (6) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (7) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (7) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (8) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (8) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (9) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (9) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (10) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (10) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (11) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (11) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (12) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (12) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (13) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (13) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (14) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (14) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (15) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (15) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (16) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (16) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (17) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (17) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (18) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (18) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (19) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (19) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (20) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (20) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (21) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (21) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (22) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (22) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (23) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (23) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (24) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (24) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (25) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (25) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (26) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (26) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (27) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (27) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (28) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (28) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
FAIL mouseover should be fired before mouseup if mousedown target is removed (29) assert_equals: mouseover should be fired before mouseup if mousedown target is removed (29) expected "mousedown@div#child,mouseover@div#parent,mouseup@div#parent" but got ""
